LPS-95509 Move tests
