+++ modules/apps/segments/segments-asah/src/test/resources/com/liferay/segments/asah/internal/client/data/binding/get-individual-segments.json
